scripts/requirements.txt
osmnx==2.0.1
psycopg2-binary==2.9.10
numpy==2.2.1
